Extract stack view component and navigator factory into separate files
